CDSTRM-1316: Add registration via NR (#114)
* CDSTRM-1316: WIP: Add registration via NR

This adds a `POST /no-auth/nr-register` endpoint to support registering via
a New Relic API key. It works, though there are a few things that still need
to be tweaked.

* CDSTRM-1316: Save NR providerInfo on user registration

This updates the `/no-auth/nr-register` endpoint so that the New Relic API key
and connection info get saved in the user object providerInfo. It also tests that
the `apiRegion` parameter is handled properly.