CDSTRM-1321: Add login by code to web (#122)
* CDSTRM-1321: Add sign-in via code to web

This enables login via code on web. This necessitates abstracting out
some of the logic from the code generation request handler into a helper
class. It also fixes an issue preventing emails from being sent when
generating a code. The page for entering the login code doesn't look
quite the same as in the IDE because implementing the separate boxes for
each number in the code like the IDE has would be much more complicated
without React and it didn't seem worth it.

* Fix input border